Make install.sh more smart
Adds some cleverness to `install.sh` as to not unnecessarily override files that don't need updating

+++ install.sh
@@ -5,9 +5,30 @@ DOTFILESDIRREL=$(dirname $0)
 cd $DOTFILESDIRREL/
 DOTFILESDIR=$(pwd -P)
 
+needs_update() {
+	if [[ ! -e "$2" ]]; then
+		return 0
+	fi
+
+	CHECKSUM1=$(md5 -q "$1")
+	CHECKSUM2=$(md5 -q "$2")
+
+	if [[ "$CHECKSUM1" != "$CHECKSUM2" ]]; then
+		return 0
+	fi
+
+	return 1
+}
+
+maybe_link() {
+	if needs_update "$1" "$2"; then
+		ln -sfv "$1" "$2"
+	fi
+}
+
 # bash
-ln -sfv "$DOTFILESDIR/bash/bashrc" "$HOME/.bashrc"
-ln -sfv "$DOTFILESDIR/bash/profile" "$HOME/.profile"
+maybe_link "$DOTFILESDIR/bash/bashrc" "$HOME/.bashrc"
+maybe_link "$DOTFILESDIR/bash/profile" "$HOME/.profile"
 
 # bin/
 if [[ -d "$HOME/.bin" ]]; then
@@ -16,18 +37,19 @@ fi
 ln -sv "$DOTFILESDIR/bin"  "$HOME/.bin"
 
 # editorconfig
-ln -sfv "$DOTFILESDIR/editorconfig/editorconfig" "$HOME/.editorconfig"
+maybe_link "$DOTFILESDIR/editorconfig/editorconfig" "$HOME/.editorconfig"
 
 # git
-ln -sfv "$DOTFILESDIR/git/gitconfig" "$HOME/.gitconfig"
-ln -sfv "$DOTFILESDIR/git/gitignore" "$HOME/.gitignore"
+maybe_link "$DOTFILESDIR/git/gitconfig" "$HOME/.gitconfig"
+maybe_link "$DOTFILESDIR/git/gitignore" "$HOME/.gitignore"
 
 # nano
-ln -sfv "$DOTFILESDIR/nano/nanorc" "$HOME/.nanorc"
+#ln -sfv "$DOTFILESDIR/nano/nanorc" "$HOME/.nanorc"
 
 # vim
-ln -sfv "$DOTFILESDIR/vim/vimrc" "$HOME/.vimrc"
-ln -sfv "$DOTFILESDIR/vim/gvimrc" "$HOME/.gvimrc"
+maybe_link "$DOTFILESDIR/vim/vimrc" "$HOME/.vimrc"
+maybe_link "$DOTFILESDIR/vim/gvimrc" "$HOME/.gvimrc"
+
 
 # Homebrew
 if [[ ! -e "/usr/local/bin/brew" ]]; then
@@ -38,8 +60,8 @@ brew bundle --global || true
 
 # Visual Studio Code
 mkdir -p "$HOME/Library/Application Support/Code/User"
-ln -sfv "$DOTFILESDIR/vscode/settings.json" "$HOME/Library/Application Support/Code/User/settings.json"
-ln -sfv "$DOTFILESDIR/vscode/keybindings.json" "$HOME/Library/Application Support/Code/User/keybindings.json"
+maybe_link "$DOTFILESDIR/vscode/settings.json" "$HOME/Library/Application Support/Code/User/settings.json"
+maybe_link "$DOTFILESDIR/vscode/keybindings.json" "$HOME/Library/Application Support/Code/User/keybindings.json"
 
 INSTALLED_EXTENSIONS="$(code --list-extensions)"
 for EXTENSION in $(cat "$DOTFILESDIR/vscode/extensions-list")
@@ -52,15 +74,17 @@ do
     fi
 done
 
-# SSH key setup.
-mkdir -p "$HOME/.ssh"
-rm -rfv "$HOME/.ssh/{id_rsa,id_rsa.pub}"
-cp "$DOTFILESDIR/ssh/id_rsa.pub" "$HOME/.ssh/"
-eval $(op signin my.1password.com j@jorgetorres.co)
-op get document "7ukxxbzsincz3bqsx6p6iwl32e" > "$HOME/.ssh/id_rsa"
-ssh-add -K
+# SSH key setup
+if needs_update "$DOTFILESDIR/ssh/id_rsa.pub" "$HOME/.ssh/id_rsa.pub"; then
+	mkdir -p "$HOME/.ssh"
+	rm -rfv "$HOME/.ssh/{id_rsa,id_rsa.pub}"
+	cp "$DOTFILESDIR/ssh/id_rsa.pub" "$HOME/.ssh/"
+	eval $(op signin my.1password.com j@jorgetorres.co)
+	op get document "7ukxxbzsincz3bqsx6p6iwl32e" > "$HOME/.ssh/id_rsa"
+	ssh-add -K
+fi
 
-# Laravel.
+# Laravel
 if [[ ! -e "/usr/local/bin/valet" ]]; then
     composer global require laravel/valet
     export PATH="$PATH:$HOME/.composer/vendor/bin"
@@ -68,10 +92,12 @@ if [[ ! -e "/usr/local/bin/valet" ]]; then
 fi
 
 # wpv
-curl https://raw.githubusercontent.com/smilingrobots/wpv/master/wpv.sh > /usr/local/bin/wpv
-chmod +x /usr/local/bin/wpv
+if [[ ! -e "/usr/local/bin/wpv" ]]; then
+	curl https://raw.githubusercontent.com/smilingrobots/wpv/master/wpv.sh > /usr/local/bin/wpv
+	chmod +x /usr/local/bin/wpv
+fi
 
-# PHP coding standards
+# PHPCS
 mkdir -p "$HOME/.phpcs"
 
 if [[ ! -d "$HOME/.phpcs/wordpress" ]]; then
